+++ .metadata/.lock_info
@@ -1,4 +1,4 @@
-#Thu May 01 18:25:39 KST 2025
+#Fri May 02 12:23:01 KST 2025
 host=DESKTOP-602G6UP
-process-id=6092
+process-id=940
 user=smhrd
