+++ .metadata/.plugins/org.eclipse.swt/EBWebView/Breadcrumbs
@@ -1,34 +1,68 @@
-0:01:32 Tab4 FinishNav8
-0:01:32 Tab4 PageLoad
-0:02:03 Microsoft.NewBrowser_Popup
-0:02:03 Microsoft.BrowserList.AddBrowser
-0:02:03 Browser5 Insert active Tab5 at 0
-0:02:03 Tab5 StartNav9 #auto_toplevel
-0:02:03 Tab5 StartNav10 #typed
-0:02:03 Tab5 FinishNav9
-0:02:03 Tab5 PageLoad
-0:02:03 Tab5 FinishNav10
-0:02:03 Tab5 PageLoad
-0:02:04 Tab5 StartNav11 #typed
-0:02:04 Tab5 FinishNav11
-0:02:04 Tab5 PageLoad
-0:02:04 Tab5 StartNav12 #typed
-0:02:04 Tab5 FinishNav12
-0:02:04 Tab5 PageLoad
-0:02:05 Microsoft.Shutdown.OnWindowClosing
-0:02:05 Microsoft.Shutdown.OnWindowClosing_MaybeClearBrowsingDataOnExit
-0:02:05 Microsoft.Shutdown.OnWindowClosing_ClearBrowsingDataOnExitNotInProgress
-0:02:05 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
-0:02:05 Tab5 RenderProcessGone
-0:02:05 Browser5 Close Tab5 at 0
-0:02:05 Tab5 WebContentsDestroyed
-0:02:05 Microsoft.Shutdown.OnWindowClosing
-0:02:05 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
-0:02:05 Widget Closed: BrowserFrame
-0:02:05 Microsoft.Shutdown.OnWindowClosing
-0:02:05 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
-0:02:05 Microsoft.Browser_Removed
-0:03:23 Microsoft.Shutdown.OnWindowClosing
+0:02:58 Tab1 StartNav22 #typed
+0:02:58 Tab1 FinishNav22
+0:02:58 Tab1 PageLoad
+0:03:07 Tab1 StartNav23 #typed
+0:03:07 Tab1 FinishNav23
+0:03:07 Tab1 PageLoad
+0:03:09 Tab1 StartNav24 #typed
+0:03:09 Tab1 FinishNav24
+0:03:09 Tab1 PageLoad
+0:03:15 Tab1 StartNav25 #typed
+0:03:15 Tab1 FinishNav25
+0:03:15 Tab1 PageLoad
+0:03:18 Tab1 StartNav26 #typed
+0:03:18 Tab1 FinishNav26
+0:03:18 Tab1 PageLoad
+0:04:28 Tab1 StartNav27 #typed
+0:04:28 Tab1 FinishNav27
+0:04:28 Tab1 PageLoad
+0:06:03 Tab1 StartNav28 #typed
+0:06:03 Tab1 FinishNav28
+0:06:03 Tab1 PageLoad
+0:06:04 Tab1 StartNav29 #typed
+0:06:04 Tab1 FinishNav29
+0:06:04 Tab1 PageLoad
+0:11:30 Tab1 StartNav30 #typed
+0:11:30 Tab1 FinishNav30
+0:11:30 Tab1 PageLoad
+0:13:05 Tab1 StartNav31 #typed
+0:13:05 Tab1 FinishNav31
+0:13:05 Tab1 PageLoad
+0:13:07 Tab1 StartNav32 #typed
+0:13:07 Tab1 FinishNav32
+0:13:07 Tab1 PageLoad
+0:13:21 Tab1 StartNav33 #typed
+0:13:21 Tab1 FinishNav33
+0:13:21 Tab1 PageLoad
+0:15:18 Tab1 StartNav34 #typed
+0:15:18 Tab1 FinishNav34
+0:15:18 Tab1 PageLoad
+0:15:33 Microsoft.Shutdown.OnWindowClosing
+0:15:33 Microsoft.Shutdown.OnWindowClosing_MaybeClearBrowsingDataOnExit
+0:15:33 Microsoft.Shutdown.OnWindowClosing_ClearBrowsingDataOnExitNotInProgress
+0:15:33 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
+0:15:33 Tab1 RenderProcessGone
+0:15:33 Browser1 Close Tab1 at 0
+0:15:33 Tab1 WebContentsDestroyed
+0:15:33 Microsoft.Shutdown.OnWindowClosing
+0:15:33 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
+0:15:33 Widget Closed: BrowserFrame
+0:15:33 Microsoft.Shutdown.OnWindowClosing
+0:15:33 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
+0:15:33 Microsoft.Last_Browser_Removed
+0:15:33 Microsoft.NewBrowser_Popup
+0:15:33 Microsoft.BrowserList.AddBrowser
+0:15:33 Browser2 Insert active Tab2 at 0
+0:15:33 Tab2 StartNav35 #auto_toplevel
+0:15:33 Tab2 StartNav36 #typed
+0:15:33 Tab2 FinishNav35
+0:15:33 Tab2 PageLoad
+0:15:33 Tab2 FinishNav36
+0:15:33 Tab2 PageLoad
+0:15:33 Tab2 StartNav37 #typed
+0:15:33 Tab2 FinishNav37
+0:15:33 Tab2 PageLoad
+0:18:09 Microsoft.Shutdown.OnWindowClosing
 0:00:00 Startup
 0:00:00 Microsoft.DeleteProfileHelper.CleanUpEphemeralProfiles
 0:00:00 Microsoft.DeleteProfileHelper.CleanUpDeletedProfiles
@@ -44,53 +78,67 @@
 0:00:00 Tab2 StartNav4 #typed
 0:00:00 Tab1 FinishNav1
 0:00:00 Tab1 PageLoad
-0:00:00 Tab2 FinishNav2
-0:00:00 Tab2 PageLoad
 0:00:00 Tab1 FinishNav3
-0:00:01 Tab2 FinishNav4
-0:00:01 Tab1 PageLoad
-0:00:01 Tab2 PageLoad
-0:02:55 Microsoft.Shutdown.OnWindowClosing
-0:00:00 Startup
-0:00:00 Microsoft.DeleteProfileHelper.CleanUpEphemeralProfiles
-0:00:00 Microsoft.DeleteProfileHelper.CleanUpDeletedProfiles
-0:00:00 Microsoft.NewBrowser_Popup
-0:00:00 Microsoft.BrowserList.AddBrowser
-0:00:00 Browser1 Insert active Tab1 at 0
-0:00:00 Tab1 StartNav1 #auto_toplevel
-0:00:00 Microsoft.NewBrowser_Popup
-0:00:00 Microsoft.BrowserList.AddBrowser
-0:00:00 Browser2 Insert active Tab2 at 0
-0:00:00 Tab2 StartNav2 #auto_toplevel
-0:00:00 Tab1 StartNav3 #typed
-0:00:00 Tab2 StartNav4 #typed
-0:00:00 Tab1 FinishNav1
-0:00:00 Tab1 PageLoad
 0:00:00 Tab2 FinishNav2
-0:00:00 Tab2 PageLoad
-0:00:00 Tab1 FinishNav3
-0:00:00 Tab2 FinishNav4
-0:00:00 Tab1 PageLoad
-0:00:00 Tab2 PageLoad
-0:03:17 Microsoft.Shutdown.OnWindowClosing
-0:00:00 Startup
-0:00:00 Microsoft.DeleteProfileHelper.CleanUpEphemeralProfiles
-0:00:00 Microsoft.DeleteProfileHelper.CleanUpDeletedProfiles
-0:00:00 Microsoft.NewBrowser_Popup
-0:00:00 Microsoft.BrowserList.AddBrowser
-0:00:00 Browser1 Insert active Tab1 at 0
-0:00:00 Tab1 StartNav1 #auto_toplevel
-0:00:00 Microsoft.NewBrowser_Popup
-0:00:00 Microsoft.BrowserList.AddBrowser
-0:00:00 Browser2 Insert active Tab2 at 0
-0:00:00 Tab2 StartNav2 #auto_toplevel
-0:00:00 Tab1 StartNav3 #typed
-0:00:00 Tab2 StartNav4 #typed
-0:00:00 Tab1 FinishNav1
-0:00:00 Tab1 PageLoad
-0:00:00 Tab2 FinishNav2
-0:00:00 Tab1 FinishNav3
 0:00:00 Tab2 PageLoad
 0:00:00 Tab2 FinishNav4
 0:00:00 Tab1 PageLoad
 0:00:00 Tab2 PageLoad
+0:09:49 Microsoft.NewBrowser_Popup
+0:09:49 Microsoft.BrowserList.AddBrowser
+0:09:49 Browser3 Insert active Tab3 at 0
+0:09:49 Tab3 StartNav5 #auto_toplevel
+0:09:49 Microsoft.NewBrowser_Popup
+0:09:49 Microsoft.BrowserList.AddBrowser
+0:09:49 Browser4 Insert active Tab4 at 0
+0:09:49 Tab4 StartNav6 #auto_toplevel
+0:09:49 Microsoft.Shutdown.OnWindowClosing
+0:09:49 Microsoft.Shutdown.OnWindowClosing_MaybeClearBrowsingDataOnExit
+0:09:49 Microsoft.Shutdown.OnWindowClosing_ClearBrowsingDataOnExitNotInProgress
+0:09:49 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
+0:09:49 Tab3 FinishNav5
+0:09:49 Tab3 RenderProcessGone
+0:09:49 Browser3 Close Tab3 at 0
+0:09:49 Tab3 WebContentsDestroyed
+0:09:49 Microsoft.Shutdown.OnWindowClosing
+0:09:49 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
+0:09:49 Widget Closed: BrowserFrame
+0:09:49 Microsoft.Shutdown.OnWindowClosing
+0:09:49 Microsoft.Shutdown.OnWindowClosingPostClearBrowsingData_NoBrowsingDataCleared
+0:09:49 Microsoft.Browser_Removed
+0:09:49 Tab4 StartNav7 #typed
+0:09:49 Tab4 FinishNav6
+0:09:49 Tab4 PageLoad
+0:09:49 Tab4 FinishNav7
+0:09:49 Tab4 PageLoad
+0:09:50 Tab4 StartNav8 #typed
+0:09:50 Tab4 FinishNav8
+0:09:50 Tab4 PageLoad
+0:09:52 Tab4 StartNav9 #typed
+0:09:52 Tab4 FinishNav9
+0:09:52 Tab4 PageLoad
+0:09:54 Tab4 StartNav10 #typed
+0:09:54 Tab4 FinishNav10
+0:09:54 Tab4 PageLoad
+0:09:54 Tab4 StartNav11 #typed
+0:09:54 Tab4 FinishNav11
+0:09:54 Tab4 PageLoad
+0:10:52 Microsoft.NewBrowser_Popup
+0:10:52 Microsoft.BrowserList.AddBrowser
+0:10:52 Browser5 Insert active Tab5 at 0
+0:10:52 Tab5 StartNav12 #auto_toplevel
+0:10:52 Tab5 StartNav13 #typed
+0:10:52 Tab5 FinishNav12
+0:10:52 Tab5 PageLoad
+0:10:52 Tab5 FinishNav13
+0:10:52 Tab5 PageLoad
+0:10:53 Tab5 StartNav14 #typed
+0:10:53 Tab5 FinishNav14
+0:10:53 Tab5 PageLoad
+0:10:54 Tab5 StartNav15 #typed
+0:10:54 Tab5 FinishNav15
+0:10:54 Tab5 PageLoad
+0:10:54 Tab5 StartNav16 #typed
+0:10:54 Tab5 FinishNav16
+0:10:54 Tab5 PageLoad
+1:44:06 Microsoft.Shutdown.OnWindowClosing
