+++ .metadata/.plugins/org.eclipse.swt/EBWebView/Default/Local Storage/leveldb/LOG
@@ -1,3 +1,3 @@
-2025/05/01-18:25:51.791 8a8 Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Local Storage\leveldb/MANIFEST-000001
-2025/05/01-18:25:51.797 8a8 Recovering log #3
-2025/05/01-18:25:51.799 8a8 Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Local Storage\leveldb/000003.log 
+2025/05/02-12:23:14.547 b28 Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Local Storage\leveldb/MANIFEST-000001
+2025/05/02-12:23:14.559 b28 Recovering log #3
+2025/05/02-12:23:14.561 b28 Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Local Storage\leveldb/000003.log
