+++ .metadata/.plugins/org.eclipse.swt/EBWebView/Default/Session Storage/LOG
@@ -1,3 +1,3 @@
-2025/05/01-18:25:51.884 8a8 Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Session Storage/MANIFEST-000001
-2025/05/01-18:25:51.885 8a8 Recovering log #3
-2025/05/01-18:25:51.888 8a8 Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Session Storage/000003.log 
+2025/05/02-12:23:14.594 b28 Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Session Storage/MANIFEST-000001
+2025/05/02-12:23:14.595 b28 Recovering log #3
+2025/05/02-12:23:14.597 b28 Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Session Storage/000003.log
