+++ .metadata/.plugins/org.eclipse.swt/EBWebView/Default/Site Characteristics Database/LOG
@@ -1,3 +1,3 @@
-2025/05/01-18:25:51.712 357c Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/MANIFEST-000001
-2025/05/01-18:25:51.713 357c Recovering log #9
-2025/05/01-18:25:51.713 357c Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/000009.log 
+2025/05/02-12:23:14.430 3b40 Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/MANIFEST-000001
+2025/05/02-12:23:14.430 3b40 Recovering log #9
+2025/05/02-12:23:14.431 3b40 Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/000009.log
