+++ .metadata/.plugins/org.eclipse.swt/EBWebView/Default/Site Characteristics Database/LOG.old
@@ -1,3 +1,3 @@
-2025/05/01-18:16:07.676 2884 Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/MANIFEST-000001
-2025/05/01-18:16:07.676 2884 Recovering log #9
-2025/05/01-18:16:07.676 2884 Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/000009.log 
+2025/05/02-09:30:16.922 28a4 Reusing MANIFEST C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/MANIFEST-000001
+2025/05/02-09:30:16.922 28a4 Recovering log #9
+2025/05/02-09:30:16.922 28a4 Reusing old log C:\Users\smhrd\Desktop\GitTest\.metadata\.plugins\org.eclipse.swt\EBWebView\Default\Site Characteristics Database/000009.log
