+++ .metadata/.plugins/org.eclipse.swt/EBWebView/Default/favorites_diagnostic.log
@@ -12,3 +12,18 @@
 2025-05-01 09:12:54.429: [INFO] OnDoneLoading sync enabled: 0
 2025-05-01 09:16:07.688: [INFO] OnDoneLoading sync enabled: 0
 2025-05-01 09:25:51.722: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 09:36:36.499: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 09:38:02.214: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 09:39:58.584: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 11:11:38.892: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 11:29:37.090: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 11:30:19.477: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 11:31:41.829: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 11:34:01.598: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 23:44:37.185: [INFO] OnDoneLoading sync enabled: 0
+2025-05-01 23:58:20.236: [INFO] OnDoneLoading sync enabled: 0
+2025-05-02 00:02:11.357: [INFO] OnDoneLoading sync enabled: 0
+2025-05-02 00:04:40.621: [INFO] OnDoneLoading sync enabled: 0
+2025-05-02 00:14:43.847: [INFO] OnDoneLoading sync enabled: 0
+2025-05-02 00:30:16.945: [INFO] OnDoneLoading sync enabled: 0
+2025-05-02 03:23:14.442: [INFO] OnDoneLoading sync enabled: 0
