+++ .metadata/.plugins/org.eclipse.swt/EBWebView/Variations
@@ -1,1 +1,1 @@
-{"user_experience_metrics.stability.exited_cleanly":false,"variations_crash_streak":0}
+{"user_experience_metrics.stability.exited_cleanly":true,"variations_crash_streak":0}
