+++ AnotherWorld/src/Audio/AudioPlayer.java
@@ -1,0 +1,40 @@
+package Audio;
+
+import java.io.File;
+import java.io.IOException;
+
+import javax.sound.sampled.AudioInputStream;
+import javax.sound.sampled.AudioSystem;
+import javax.sound.sampled.Clip;
+import javax.sound.sampled.LineUnavailableException;
+import javax.sound.sampled.UnsupportedAudioFileException;
+
+public class AudioPlayer {
+	private Clip clip;
+
+    public void load(String filePath) throws UnsupportedAudioFileException, IOException, LineUnavailableException {
+        File audioFile = new File(filePath);
+        AudioInputStream audioStream = AudioSystem.getAudioInputStream(audioFile);
+        clip = AudioSystem.getClip();
+        clip.open(audioStream);
+    }
+
+    public void play() {
+        if (clip != null) {
+            clip.start();
+        }
+    }
+
+    public void stop() {
+        if (clip != null && clip.isRunning()) {
+            clip.stop();
+        }
+    }
+
+    public void close() {
+        if (clip != null) {
+            clip.close();
+        }
+    }
+
+}
