--- AnotherWorld/src/player/QuizService.java
@@ -1,5 +1,0 @@
-package player;
-
-public class QuizService {
-
-}
